# src/main/java/com/inventorilab/dto/response/KategoriResponse.java
package com.inventorilab.dto.response;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class KategoriResponse {
    private Long idKategori;
    private String namaKategori;
    private String deskripsi;
}